+++ src/main/java/com/bulletphysics/collision/broadphase/AxisSweep3Internal.java
@@ -31,7 +31,7 @@ import com.bulletphysics.collision.broadphase.AxisSweep3Internal.Handle;
 import com.bulletphysics.linearmath.MiscUtil;
 import com.bulletphysics.linearmath.VectorUtil;
 import com.bulletphysics.util.ObjectArrayList;
-import javax.vecmath.Vector3f;
+import javax.vecmath.Vector3d;
 
 /**
  * AxisSweep3Internal is an internal base class that implements sweep and prune.
@@ -44,10 +44,10 @@ public abstract class AxisSweep3Internal extends BroadphaseInterface {
 	protected int bpHandleMask;
 	protected int handleSentinel;
 	
-	protected final Vector3f worldAabbMin = new Vector3f(); // overall system bounds
-	protected final Vector3f worldAabbMax = new Vector3f(); // overall system bounds
+	protected final Vector3d worldAabbMin = new Vector3d(); // overall system bounds
+	protected final Vector3d worldAabbMax = new Vector3d(); // overall system bounds
 
-	protected final Vector3f quantize = new Vector3f();     // scaling factor for quantization
+	protected final Vector3d quantize = new Vector3d();     // scaling factor for quantization
 
 	protected int numHandles;                               // number of active handles
 	protected int maxHandles;                               // max number of handles
@@ -68,7 +68,7 @@ public abstract class AxisSweep3Internal extends BroadphaseInterface {
 	// JAVA NOTE: added
 	protected int mask;
 	
-	AxisSweep3Internal(Vector3f worldAabbMin, Vector3f worldAabbMax, int handleMask, int handleSentinel, int userMaxHandles/* = 16384*/, OverlappingPairCache pairCache/*=0*/) {
+	AxisSweep3Internal(Vector3d worldAabbMin, Vector3d worldAabbMax, int handleMask, int handleSentinel, int userMaxHandles/* = 16384*/, OverlappingPairCache pairCache/*=0*/) {
 		this.bpHandleMask = handleMask;
 		this.handleSentinel = handleSentinel;
 		this.pairCache = pairCache;
@@ -86,7 +86,7 @@ public abstract class AxisSweep3Internal extends BroadphaseInterface {
 		this.worldAabbMin.set(worldAabbMin);
 		this.worldAabbMax.set(worldAabbMax);
 
-		Vector3f aabbSize = new Vector3f();
+		Vector3d aabbSize = new Vector3d();
 		aabbSize.sub(this.worldAabbMax, this.worldAabbMin);
 
 		int maxInt = this.handleSentinel;
@@ -190,13 +190,13 @@ public abstract class AxisSweep3Internal extends BroadphaseInterface {
 	//void debugPrintAxis(int axis,bool checkCardinality=true);
 	//#endif //DEBUG_BROADPHASE
 
-	protected void quantize(int[] out, Vector3f point, int isMax) {
-		Vector3f clampedPoint = new Vector3f(point);
+	protected void quantize(int[] out, Vector3d point, int isMax) {
+		Vector3d clampedPoint = new Vector3d(point);
 
 		VectorUtil.setMax(clampedPoint, worldAabbMin);
 		VectorUtil.setMin(clampedPoint, worldAabbMax);
 
-		Vector3f v = new Vector3f();
+		Vector3d v = new Vector3d();
 		v.sub(clampedPoint, worldAabbMin);
 		VectorUtil.mul(v, v, quantize);
 
@@ -442,7 +442,7 @@ public abstract class AxisSweep3Internal extends BroadphaseInterface {
 		}
 	}
 	
-	public int addHandle(Vector3f aabbMin, Vector3f aabbMax, Object pOwner, short collisionFilterGroup, short collisionFilterMask, Dispatcher dispatcher, Object multiSapProxy) {
+	public int addHandle(Vector3d aabbMin, Vector3d aabbMax, Object pOwner, short collisionFilterGroup, short collisionFilterMask, Dispatcher dispatcher, Object multiSapProxy) {
 		// quantize the bounds
 		int[] min = new int[3], max = new int[3];
 		quantize(min, aabbMin, 0);
@@ -534,7 +534,7 @@ public abstract class AxisSweep3Internal extends BroadphaseInterface {
 		freeHandle(handle);
 	}
 	
-	public void updateHandle(int handle, Vector3f aabbMin, Vector3f aabbMax, Dispatcher dispatcher) {
+	public void updateHandle(int handle, Vector3d aabbMin, Vector3d aabbMax, Dispatcher dispatcher) {
 		Handle pHandle = getHandle(handle);
 
 		// quantize the new bounds
@@ -580,7 +580,7 @@ public abstract class AxisSweep3Internal extends BroadphaseInterface {
 	//public void processAllOverlappingPairs(OverlapCallback callback) {
 	//}
 	
-	public BroadphaseProxy createProxy(Vector3f aabbMin, Vector3f aabbMax, BroadphaseNativeType shapeType, Object userPtr, short collisionFilterGroup, short collisionFilterMask, Dispatcher dispatcher, Object multiSapProxy) {
+	public BroadphaseProxy createProxy(Vector3d aabbMin, Vector3d aabbMax, BroadphaseNativeType shapeType, Object userPtr, short collisionFilterGroup, short collisionFilterMask, Dispatcher dispatcher, Object multiSapProxy) {
 		int handleId = addHandle(aabbMin, aabbMax, userPtr, collisionFilterGroup, collisionFilterMask, dispatcher, multiSapProxy);
 
 		Handle handle = getHandle(handleId);
@@ -593,7 +593,7 @@ public abstract class AxisSweep3Internal extends BroadphaseInterface {
 		removeHandle(handle.uniqueId, dispatcher);
 	}
 
-	public void setAabb(BroadphaseProxy proxy, Vector3f aabbMin, Vector3f aabbMax, Dispatcher dispatcher) {
+	public void setAabb(BroadphaseProxy proxy, Vector3d aabbMin, Vector3d aabbMax, Dispatcher dispatcher) {
 		Handle handle = (Handle) proxy;
 		updateHandle(handle.uniqueId, aabbMin, aabbMax, dispatcher);
 	}
@@ -627,7 +627,7 @@ public abstract class AxisSweep3Internal extends BroadphaseInterface {
 	
 	// getAabb returns the axis aligned bounding box in the 'global' coordinate frame
 	// will add some transform later
-	public void getBroadphaseAabb(Vector3f aabbMin, Vector3f aabbMax) {
+	public void getBroadphaseAabb(Vector3d aabbMin, Vector3d aabbMax) {
 		aabbMin.set(worldAabbMin);
 		aabbMax.set(worldAabbMax);
 	}
